Adjust hue saturation value for filtering red
Red for hue is from ~0-60, so I adjusted to 0-80. I'm adjusting the others to try and work in the bright lighting of the lab. A matte ball works better in these circumstances. depending on the area we do the tests we may need to set these parameters differently. I have a program which allows us to slide and detect the values so it's just a matter of setting things up and testing in the environment.